+++ resources/views/users/index.blade.php
@@ -68,42 +68,45 @@ jQuery(document).ready(function($) {
                                     </td>
                                     <td>{{ strtoupper(implode(', ', json_decode(json_encode($user->getRoleNames()), true))) }}</td>
                                     <td>
-                                        @if($user->deleted_at !== NULL)
-                                            @if(auth()->user()->can('delete_hard_users'))
-                                                <button type="button" class="btn btn-sm btn-success" data-toggle="modal" data-target="#userModal"
-                                                    data-title="Are you sure you want to activate the user {{ $user->name }}?"
-                                                    data-body="All user data will be restored."
-                                                    data-form="{{ route('users.restore', $user->id) }}"
-                                                    data-method="POST"
-                                                    data-class="btn btn-sm btn-success"
-                                                    data-action="Activate">Activate</button>
-                                                <button type="button" class="btn btn-sm btn-danger" data-toggle="modal" data-target="#userModal"
-                                                    data-title="Are you sure you want to delete the user {{ $user->name }}?"
-                                                    data-body="WARNING: All user data will be deleted. This action can not be undone."
-                                                    data-form="{{ route('users.destroy', $user->id) }}"
-                                                    data-method="DELETE"
-                                                    data-class="btn btn-sm btn-danger"
-                                                    data-action="Delete">Delete</button>
-                                            @endif
-                                        @else
-                                            @if(auth()->user()->can('delete_hard_users'))
-                                                <button type="button" class="btn btn-sm btn-warning" data-toggle="modal" data-target="#userModal"
-                                                    data-title="Are you sure you want to deactivate the user {{ $user->name }}?"
-                                                    data-body="The data of the user will remain in the application, but the users that depend on it will not be able to access the data."
-                                                    data-form="{{ route('users.delete', $user) }}"
-                                                    data-method="DELETE"
-                                                    data-class="btn btn-sm btn-warning"
-                                                    data-action="Deactivate">Deactivate</button>
+                                        @if(\Auth::user()->id !== $user->id)
+                                            @if($user->deleted_at !== NULL)
+                                                @if(auth()->user()->can('delete_hard_users'))
+                                                    <button type="button" class="btn btn-sm btn-success" data-toggle="modal" data-target="#userModal"
+                                                        data-title="Are you sure you want to activate the user {{ $user->name }}?"
+                                                        data-body="All user data will be restored."
+                                                        data-form="{{ route('users.restore', $user->id) }}"
+                                                        data-method="POST"
+                                                        data-class="btn btn-sm btn-success"
+                                                        data-action="Activate">Activate</button>
+                                                    <button type="button" class="btn btn-sm btn-danger" data-toggle="modal" data-target="#userModal"
+                                                        data-title="Are you sure you want to delete the user {{ $user->name }}?"
+                                                        data-body="WARNING: All user data will be deleted. This action can not be undone."
+                                                        data-form="{{ route('users.destroy', $user->id) }}"
+                                                        data-method="DELETE"
+                                                        data-class="btn btn-sm btn-danger"
+                                                        data-action="Delete">Delete</button>
+                                                @endif
                                             @else
-                                                <button type="button" class="btn btn-sm btn-warning" data-toggle="modal" data-target="#userModal"
-                                                    data-title="Are you sure you want to deactivate the user {{ $user->name }}?"
-                                                    data-body="The data of the user will remain in the application, but the users that depend on it will not be able to access the data."
-                                                    data-form="{{ route('users.delete', $user) }}"
-                                                    data-method="DELETE"
-                                                    data-class="btn btn-sm btn-warning"
-                                                    data-action="Delete">Delete</button>
+                                                @if(auth()->user()->can('delete_hard_users'))
+                                                    <button type="button" class="btn btn-sm btn-warning" data-toggle="modal" data-target="#userModal"
+                                                        data-title="Are you sure you want to deactivate the user {{ $user->name }}?"
+                                                        data-body="The data of the user will remain in the application, but the users that depend on it will not be able to access the data."
+                                                        data-form="{{ route('users.delete', $user) }}"
+                                                        data-method="DELETE"
+                                                        data-class="btn btn-sm btn-warning"
+                                                        data-action="Deactivate">Deactivate</button>
+                                                @else
+                                                    <button type="button" class="btn btn-sm btn-warning" data-toggle="modal" data-target="#userModal"
+                                                        data-title="Are you sure you want to deactivate the user {{ $user->name }}?"
+                                                        data-body="The data of the user will remain in the application, but the users that depend on it will not be able to access the data."
+                                                        data-form="{{ route('users.delete', $user) }}"
+                                                        data-method="DELETE"
+                                                        data-class="btn btn-sm btn-warning"
+                                                        data-action="Delete">Delete</button>
+                                                @endif
                                             @endif
                                         @endif
+
                                     </td>
                                 </tr>
                             @empty
